agent/requirements.txt
pyperclip==1.8.2
python-socketio[client]==5.10.0
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.7.0
